turn countLanguagesFromVisibleActivities into a selector
