+++ lib/charts.js
@@ -76,38 +76,14 @@ export const getActivityTimeChartData = (activities, moment) => {
 }
 
 
-export const countLanguagesFromActivities = (activities, languageMap, actorMap) => {
-  return activities.reduce((languages, activity) => {
-    const rLanguages = languages;
-    const activityLanguages = actorMap[activity.actor].languages.map(lang => languageMap[lang]);
-    if (activityLanguages && activityLanguages.length) {
-      activityLanguages.forEach((lang) => {
-        if (lang.name in rLanguages) {
-          rLanguages[lang.name].count += 1;
-        } else {
-          rLanguages[lang.name] = {
-            count: 1,
-            color: lang.color,
-            name: lang.name,
-          };
-        }
-      });
-    }
-    return rLanguages;
-  }, {});
-}
-
-
-export const getActivityLanguagePieData = (activities, languageMap, actorMap) => {
+export const getActivityLanguagePieData = (languageCount) => {
   /**
    * Return activity data structured for a line chart based on activities.
    *
    * @param activities {Array, <Object>}
    */
-  let l = countLanguagesFromActivities(activities, languageMap, actorMap);
-
   // flatten
-  l = Object.keys(l).map(key => l[key]);
+  const l = Object.keys(languageCount).map(key => languageCount[key]);
   return {
     labels: l.map(lang => lang.name),
     datasets: [{
