+++ lib/storage.js
@@ -137,3 +137,25 @@ export const selectVisibleActivityIds = (selectedLanguages, selectedActors, acti
   }
   return filteredActivities.map(activity => activity.id);
 };
+
+
+export const countLanguagesFromActivities = (activities, languageMap, actorMap) => {
+  return activities.reduce((languages, activity) => {
+    const rLanguages = languages;
+    const activityLanguages = actorMap[activity.actor].languages.map(lang => languageMap[lang]);
+    if (activityLanguages && activityLanguages.length) {
+      activityLanguages.forEach((lang) => {
+        if (lang.name in rLanguages) {
+          rLanguages[lang.name].count += 1;
+        } else {
+          rLanguages[lang.name] = {
+            count: 1,
+            color: lang.color,
+            name: lang.name,
+          };
+        }
+      });
+    }
+    return rLanguages;
+  }, {});
+}
